requirements.txt
grpcio
grpcio-tools
textblob
flask
flask-cors